script clean up
